undid some changes and added abstract Policy
