+++ hbaselines/base_policies/__init__.py
@@ -1,8 +1,10 @@
 """Init file for the base policies submodule."""
+from hbaselines.base_policies.policy import Policy
 from hbaselines.base_policies.actor_critic import ActorCriticPolicy
 from hbaselines.base_policies.imitation import ImitationLearningPolicy
 
 __all__ = [
+    "Policy",
     "ActorCriticPolicy",
     "ImitationLearningPolicy"
 ]
