+++ hbaselines/fcnet/td3.py
@@ -93,15 +93,15 @@ class FeedForwardPolicy(ActorCriticPolicy):
                  ob_space,
                  ac_space,
                  co_space,
-                 verbose,
-                 model_params,
                  buffer_size,
                  batch_size,
                  actor_lr,
                  critic_lr,
+                 verbose,
                  tau,
                  gamma,
                  use_huber,
+                 model_params,
                  noise,
                  target_policy_noise,
                  target_noise_clip,
@@ -118,11 +118,6 @@ class FeedForwardPolicy(ActorCriticPolicy):
             the action space of the environment
         co_space : gym.spaces.*
             the context space of the environment
-        verbose : int
-            the verbosity level: 0 none, 1 training information, 2 tensorflow
-            debug
-        model_params : dict
-            dictionary of model-specific parameters. See parent class.
         buffer_size : int
             the max number of transitions to store
         batch_size : int
@@ -131,6 +126,9 @@ class FeedForwardPolicy(ActorCriticPolicy):
             actor learning rate
         critic_lr : float
             critic learning rate
+        verbose : int
+            the verbosity level: 0 none, 1 training information, 2 tensorflow
+            debug
         tau : float
             target update rate
         gamma : float
@@ -139,6 +137,8 @@ class FeedForwardPolicy(ActorCriticPolicy):
             specifies whether to use the huber distance function as the loss
             for the critic. If set to False, the mean-squared error metric is
             used instead
+        model_params : dict
+            dictionary of model-specific parameters. See parent class.
         noise : float
             scaling term to the range of the action space, that is subsequently
             used as the standard deviation of Gaussian noise added to the
@@ -156,17 +156,16 @@ class FeedForwardPolicy(ActorCriticPolicy):
             ob_space=ob_space,
             ac_space=ac_space,
             co_space=co_space,
+            buffer_size=buffer_size,
+            batch_size=batch_size,
+            actor_lr=actor_lr,
+            critic_lr=critic_lr,
             verbose=verbose,
+            tau=tau,
+            gamma=gamma,
+            use_huber=use_huber,
             model_params=model_params,
         )
-
-        self.buffer_size = buffer_size
-        self.batch_size = batch_size
-        self.actor_lr = actor_lr
-        self.critic_lr = critic_lr
-        self.tau = tau
-        self.gamma = gamma
-        self.use_huber = use_huber
 
         # action magnitudes
         ac_mag = 0.5 * (ac_space.high - ac_space.low)
