+++ hbaselines/goal_conditioned/base.py
@@ -48,22 +48,6 @@ class GoalConditionedPolicy(ActorCriticPolicy):
 
     Attributes
     ----------
-    buffer_size : int
-        the max number of transitions to store
-    batch_size : int
-        SGD batch size
-    actor_lr : float
-        actor learning rate
-    critic_lr : float
-        critic learning rate
-    tau : float
-        target update rate
-    gamma : float
-        discount factor
-    use_huber : bool
-        specifies whether to use the huber distance function as the loss for
-        the critic. If set to False, the mean-squared error metric is used
-        instead
     num_levels : int
         number of levels within the hierarchy. Must be greater than 1. Two
         levels correspond to a Manager/Worker paradigm.
@@ -163,11 +147,6 @@ class GoalConditionedPolicy(ActorCriticPolicy):
             the action space of the environment
         co_space : gym.spaces.*
             the context space of the environment
-        verbose : int
-            the verbosity level: 0 none, 1 training information, 2 tensorflow
-            debug
-        model_params : dict
-            dictionary of model-specific parameters. See parent class.
         buffer_size : int
             the max number of transitions to store
         batch_size : int
@@ -176,6 +155,9 @@ class GoalConditionedPolicy(ActorCriticPolicy):
             actor learning rate
         critic_lr : float
             critic learning rate
+        verbose : int
+            the verbosity level: 0 none, 1 training information, 2 tensorflow
+            debug
         tau : float
             target update rate
         gamma : float
@@ -184,6 +166,8 @@ class GoalConditionedPolicy(ActorCriticPolicy):
             specifies whether to use the huber distance function as the loss
             for the critic. If set to False, the mean-squared error metric is
             used instead
+        model_params : dict
+            dictionary of model-specific parameters. See parent class.
         num_levels : int
             number of levels within the hierarchy. Must be greater than 1. Two
             levels correspond to a Manager/Worker paradigm.
@@ -244,19 +228,19 @@ class GoalConditionedPolicy(ActorCriticPolicy):
             ob_space=ob_space,
             ac_space=ac_space,
             co_space=co_space,
+            buffer_size=buffer_size,
+            batch_size=batch_size,
+            actor_lr=actor_lr,
+            critic_lr=critic_lr,
             verbose=verbose,
+            tau=tau,
+            gamma=gamma,
+            use_huber=use_huber,
             model_params=model_params,
         )
 
         assert num_levels >= 2, "num_levels must be greater than or equal to 2"
 
-        self.buffer_size = buffer_size
-        self.batch_size = batch_size
-        self.actor_lr = actor_lr
-        self.critic_lr = critic_lr
-        self.tau = tau
-        self.gamma = gamma
-        self.use_huber = use_huber
         self.num_levels = num_levels
         self.meta_period = meta_period
         self.intrinsic_reward_type = intrinsic_reward_type
