+++ hbaselines/multiagent/h_td3.py
@@ -149,7 +149,14 @@ class MultiGoalConditionedPolicy(BasePolicy):
             ob_space=ob_space,
             ac_space=ac_space,
             co_space=co_space,
+            buffer_size=buffer_size,
+            batch_size=batch_size,
+            actor_lr=actor_lr,
+            critic_lr=critic_lr,
             verbose=verbose,
+            tau=tau,
+            gamma=gamma,
+            use_huber=use_huber,
             model_params=model_params,
             shared=shared,
             maddpg=maddpg,
@@ -158,13 +165,6 @@ class MultiGoalConditionedPolicy(BasePolicy):
             base_policy=GoalConditionedPolicy,
             scope=scope,
             additional_params=dict(
-                buffer_size=buffer_size,
-                batch_size=batch_size,
-                actor_lr=actor_lr,
-                critic_lr=critic_lr,
-                tau=tau,
-                gamma=gamma,
-                use_huber=use_huber,
                 noise=noise,
                 target_policy_noise=target_policy_noise,
                 target_noise_clip=target_noise_clip,
