+++ hbaselines/multiagent/sac.py
@@ -33,22 +33,6 @@ class MultiFeedForwardPolicy(BasePolicy):
 
     Attributes
     ----------
-    buffer_size : int
-        the max number of transitions to store
-    batch_size : int
-        SGD batch size
-    actor_lr : float
-        actor learning rate
-    critic_lr : float
-        critic learning rate
-    tau : float
-        target update rate
-    gamma : float
-        discount factor
-    use_huber : bool
-        specifies whether to use the huber distance function as the loss for
-        the critic. If set to False, the mean-squared error metric is used
-        instead
     target_entropy : float
         target entropy used when learning the entropy coefficient
     replay_buffer : MultiReplayBuffer or SharedReplayBuffer
@@ -116,15 +100,15 @@ class MultiFeedForwardPolicy(BasePolicy):
                  ob_space,
                  ac_space,
                  co_space,
-                 verbose,
-                 model_params,
                  buffer_size,
                  batch_size,
                  actor_lr,
                  critic_lr,
+                 verbose,
                  tau,
                  gamma,
                  use_huber,
+                 model_params,
                  target_entropy,
                  shared,
                  maddpg,
@@ -143,11 +127,6 @@ class MultiFeedForwardPolicy(BasePolicy):
             the action space of the environment
         co_space : gym.spaces.*
             the context space of the environment
-        verbose : int
-            the verbosity level: 0 none, 1 training information, 2 tensorflow
-            debug
-        model_params : dict
-            dictionary of model-specific parameters. See parent class.
         buffer_size : int
             the max number of transitions to store
         batch_size : int
@@ -156,6 +135,9 @@ class MultiFeedForwardPolicy(BasePolicy):
             actor learning rate
         critic_lr : float
             critic learning rate
+        verbose : int
+            the verbosity level: 0 none, 1 training information, 2 tensorflow
+            debug
         tau : float
             target update rate
         gamma : float
@@ -164,6 +146,8 @@ class MultiFeedForwardPolicy(BasePolicy):
             specifies whether to use the huber distance function as the loss
             for the critic. If set to False, the mean-squared error metric is
             used instead
+        model_params : dict
+            dictionary of model-specific parameters. See parent class.
         target_entropy : float
             target entropy used when learning the entropy coefficient. If set
             to None, a heuristic value is used.
@@ -182,14 +166,6 @@ class MultiFeedForwardPolicy(BasePolicy):
         scope : str
             an upper-level scope term. Used by policies that call this one.
         """
-        self.buffer_size = buffer_size
-        self.batch_size = batch_size
-        self.actor_lr = actor_lr
-        self.critic_lr = critic_lr
-        self.tau = tau
-        self.gamma = gamma
-        self.use_huber = use_huber
-
         # Instantiate a few terms (needed if MADDPG is used).
         if target_entropy is None:
             if isinstance(ac_space, dict):
@@ -249,7 +225,14 @@ class MultiFeedForwardPolicy(BasePolicy):
             ob_space=ob_space,
             ac_space=ac_space,
             co_space=co_space,
+            buffer_size=buffer_size,
+            batch_size=batch_size,
+            actor_lr=actor_lr,
+            critic_lr=critic_lr,
             verbose=verbose,
+            tau=tau,
+            gamma=gamma,
+            use_huber=use_huber,
             model_params=model_params,
             shared=shared,
             maddpg=maddpg,
@@ -258,13 +241,6 @@ class MultiFeedForwardPolicy(BasePolicy):
             base_policy=FeedForwardPolicy,
             scope=scope,
             additional_params=dict(
-                buffer_size=buffer_size,
-                batch_size=batch_size,
-                actor_lr=actor_lr,
-                critic_lr=critic_lr,
-                tau=tau,
-                gamma=gamma,
-                use_huber=use_huber,
                 target_entropy=target_entropy,
             ),
         )
