+++ hbaselines/multiagent/td3.py
@@ -25,22 +25,6 @@ class MultiFeedForwardPolicy(BasePolicy):
 
     Attributes
     ----------
-    buffer_size : int
-        the max number of transitions to store
-    batch_size : int
-        SGD batch size
-    actor_lr : float
-        actor learning rate
-    critic_lr : float
-        critic learning rate
-    tau : float
-        target update rate
-    gamma : float
-        discount factor
-    use_huber : bool
-        specifies whether to use the huber distance function as the loss for
-        the critic. If set to False, the mean-squared error metric is used
-        instead
     noise : float
         scaling term to the range of the action space, that is subsequently
         used as the standard deviation of Gaussian noise added to the action if
@@ -95,15 +79,15 @@ class MultiFeedForwardPolicy(BasePolicy):
                  ob_space,
                  ac_space,
                  co_space,
-                 verbose,
-                 model_params,
                  buffer_size,
                  batch_size,
                  actor_lr,
                  critic_lr,
+                 verbose,
                  tau,
                  gamma,
                  use_huber,
+                 model_params,
                  noise,
                  target_policy_noise,
                  target_noise_clip,
@@ -124,11 +108,6 @@ class MultiFeedForwardPolicy(BasePolicy):
             the action space of the environment
         co_space : gym.spaces.*
             the context space of the environment
-        verbose : int
-            the verbosity level: 0 none, 1 training information, 2 tensorflow
-            debug
-        model_params : dict
-            dictionary of model-specific parameters. See parent class.
         buffer_size : int
             the max number of transitions to store
         batch_size : int
@@ -137,6 +116,9 @@ class MultiFeedForwardPolicy(BasePolicy):
             actor learning rate
         critic_lr : float
             critic learning rate
+        verbose : int
+            the verbosity level: 0 none, 1 training information, 2 tensorflow
+            debug
         tau : float
             target update rate
         gamma : float
@@ -145,6 +127,8 @@ class MultiFeedForwardPolicy(BasePolicy):
             specifies whether to use the huber distance function as the loss
             for the critic. If set to False, the mean-squared error metric is
             used instead
+        model_params : dict
+            dictionary of model-specific parameters. See parent class.
         noise : float
             scaling term to the range of the action space, that is subsequently
             used as the standard deviation of Gaussian noise added to the
@@ -169,14 +153,6 @@ class MultiFeedForwardPolicy(BasePolicy):
         scope : str
             an upper-level scope term. Used by policies that call this one.
         """
-        self.buffer_size = buffer_size
-        self.batch_size = batch_size
-        self.actor_lr = actor_lr
-        self.critic_lr = critic_lr
-        self.tau = tau
-        self.gamma = gamma
-        self.use_huber = use_huber
-
         # Instantiate a few terms (needed if MADDPG is used).
         if shared:
             # action magnitudes
@@ -224,7 +200,14 @@ class MultiFeedForwardPolicy(BasePolicy):
             ob_space=ob_space,
             ac_space=ac_space,
             co_space=co_space,
+            buffer_size=buffer_size,
+            batch_size=batch_size,
+            actor_lr=actor_lr,
+            critic_lr=critic_lr,
             verbose=verbose,
+            tau=tau,
+            gamma=gamma,
+            use_huber=use_huber,
             model_params=model_params,
             shared=shared,
             maddpg=maddpg,
@@ -233,13 +216,6 @@ class MultiFeedForwardPolicy(BasePolicy):
             base_policy=FeedForwardPolicy,
             scope=scope,
             additional_params=dict(
-                buffer_size=buffer_size,
-                batch_size=batch_size,
-                actor_lr=actor_lr,
-                critic_lr=critic_lr,
-                tau=tau,
-                gamma=gamma,
-                use_huber=use_huber,
                 noise=noise,
                 target_policy_noise=target_policy_noise,
                 target_noise_clip=target_noise_clip,
